Add universal variable notifier based off of shared memory
This is a conceptually more straight-forward solution for synchronizing
data between multiple fish instances, using the straightforward shared
memory approach.

I'm not sure how the SIGIO notifier is actually "notifying" since it
just returns true/false when polled, but if it is desirable to use the
notifier in a non-polling manner, the `block()` method will sleep the
calling thread until a uvar change is detected. If it's desired to use
it in an asynchronous context with an event loop (e.g. with
select/poll/epoll), it is trivial to add an anonymous pipe and use the
self-pipe trick to provide an fd that can be given to the kernel to wait
on by calling `notifier.block()` in a new thread and writing (then
immediately reading) a byte to/from the pipe.

This notifier has been tested on Linux, FreeBSD, WSL, and macOS and
passes all the tests. There is currently no `#ifdef` for selecting the
strategy, suggestions are welcome.

The code currently uses the GCC atomic extensions (CMake patched
accordingly) to atomically read/write the one- to four-byte fields in
the shared memory. This may be overkill: presumably the pointer nature
of the shared memory object would suffice to prevent the compiler from
making assumptions that would lead to stale reads; the addition of a
`volatile` modifier to the storage class of either the SHM pointer or
the individual fields would definitely suffice, but that's apparently no
longer "in vogue" (suggestions again welcome).