Revert "Attempt to simplify how completions get presented in the pager"
The pager cleanup missed that the existing token could already include active (as in unescaped) expansions, and just escaped them all.

This means things like `ls ~/<TAB>` would escape the `~`, which is obviously wrong and makes it awkward to use.

This reverts commit b38a23a46dfac9fdaf453cced61683765790528c.

I fully expect that we'll try again, but there's no use in keeping master broken while that happens.

Fixes #7526.